This set of functions extract precipitation data from locally stored AgERA5 data

This set of functions extract precipitation data from locally stored AgERA5 data

Usage

get_prec(...)

# S3 method for period
get_prec(.start_date, .end_date, .lon, .lat, .agera5_folder)

# S3 method for dataset
get_prec(
  .trial_dataset = NULL,
  .start_date = "pdate",
  .end_date = "hdate",
  .lon = "lon",
  .lat = "lat",
  .agera5_folder
)

Arguments

.start_date

Date or character to be coerced as Date The starting date for the period to extract

.end_date

Date or character to be coerced as Date The end date for the period to extract

.lon

numeric Longitude for the point of interest

.lat

numeric Latitude for the point of interest

.agera5_folder

Folder in the local system where the agera5 data is stored. Usually the root folder will work as the function will search recursively

.date

a date or character representing the date of the data point to be extracted

Methods (by class)

period: Get precipitation data for one location for a provided time period

dataset: Iterates across a data set to extract all required data points